+++ src/Http/Requests/CreateShippingMethod.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Contains the CreateShippingMethod class.
+ *
+ * @copyright   Copyright (c) 2023 Vanilo UG
+ * @author      Attila Fulop
+ * @license     MIT
+ * @since       2023-01-25
+ *
+ */
+
+namespace Vanilo\Admin\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Vanilo\Admin\Contracts\Requests\CreateShippingMethod as CreateShippingMethodContract;
+
+class CreateShippingMethod extends FormRequest implements CreateShippingMethodContract
+{
+    public function rules()
+    {
+        return [
+            'name' => 'required|min:2|max:255',
+            'carrier_id' => 'required|exists:carriers,id',
+            'configuration' => 'sometimes|json',
+            'is_active' => 'sometimes|boolean',
+        ];
+    }
+
+    public function authorize()
+    {
+        return true;
+    }
+}